+++ rustfmt.toml
@@ -1,0 +1,1 @@
+wrap_comments = true
